Replace min-active-replica instance-drop guard rail with a live-instance check
The DELETE instance guard rail previously ran MinActiveReplicaGuardrailRule,
which scanned every resource's ideal state / external view to reject a drop that
would breach min active replicas. On the actual delete path that check is both
expensive and near-silent: an instance can only be dropped once it is already
offline, at which point its at-risk partitions are skipped and the rule passes
regardless of replication state. The meaningful precondition for dropping an
instance is simply that it is no longer live.

Replace it with LiveInstanceGuardrailRule, which reads only the target's single
LIVEINSTANCES znode and blocks (or, for dryRun, reports) the drop while the
participant is still connected. This mirrors the authoritative check already in
ZKHelixAdmin.dropInstance but surfaces it as a structured pre-flight verdict and
makes dryRun truthful, at negligible ZK read cost. The admin-layer check is
retained as the enforcer for non-REST callers and to close the TOCTOU gap.

- Add LiveInstanceGuardrailRule (+ unit test)
- Wire it into PerInstanceAccessor.deleteInstance in place of the min-active rule
- Update TestPerInstanceAccessor delete guard-rail coverage
- Remove MinActiveReplicaGuardrailRule and its unit test

The shared InstanceValidationUtil.siblingNodesActiveReplicaCheckWithDetails and
MinActiveReplicaCheckResult helpers are left intact; they remain in use by
MaintenanceManagementService.
