+++ helix-rest/src/main/java/org/apache/helix/rest/server/resources/helix/PerInstanceAccessor.java
@@ -54,7 +54,7 @@ import org.apache.helix.constants.InstanceDrainExclusionType;
 import org.apache.helix.constants.InstanceConstants;
 import org.apache.helix.guardrail.GuardrailContext;
 import org.apache.helix.guardrail.GuardrailPipeline;
-import org.apache.helix.guardrail.rules.MinActiveReplicaGuardrailRule;
+import org.apache.helix.guardrail.rules.LiveInstanceGuardrailRule;
 import org.apache.helix.manager.zk.ZKHelixAdmin;
 import org.apache.helix.manager.zk.ZKHelixDataAccessor;
 import org.apache.helix.manager.zk.ZkBaseDataAccessor;
@@ -648,13 +648,15 @@ public class PerInstanceAccessor extends AbstractHelixResource {
       @PathParam("instanceName") String instanceName,
       @DefaultValue("false") @QueryParam("force") boolean force,
       @DefaultValue("false") @QueryParam("dryRun") boolean dryRun) {
-    // Guard rail: block (or simulate) a drop that would violate any resource's minimum active
-    // replica count. force=true overrides; dryRun=true only reports the verdict without dropping.
+    // Guard rail: block (or simulate) dropping an instance that is still live -- its participant is
+    // connected, so its LIVEINSTANCES znode is present. force=true overrides this verdict, though
+    // the admin layer (ZKHelixAdmin.dropInstance) still rejects dropping a live instance;
+    // dryRun=true only reports the verdict without dropping.
     GuardrailContext context = GuardrailContext.newBuilder(clusterId)
         .dataAccessor(getDataAccssor(clusterId))
         .instanceName(instanceName)
         .build();
-    GuardrailPipeline pipeline = new GuardrailPipeline(new MinActiveReplicaGuardrailRule());
+    GuardrailPipeline pipeline = new GuardrailPipeline(new LiveInstanceGuardrailRule());
     Optional<Response> preflightResponse = preflight(pipeline, context, force, dryRun);
     if (preflightResponse.isPresent()) {
       return preflightResponse.get();
