+++ helix-rest/src/test/java/org/apache/helix/rest/server/TestPerInstanceAccessor.java
@@ -42,7 +42,7 @@ import org.apache.helix.HelixDefinedState;
 import org.apache.helix.HelixException;
 import org.apache.helix.TestHelper;
 import org.apache.helix.constants.InstanceConstants;
-import org.apache.helix.guardrail.rules.MinActiveReplicaGuardrailRule;
+import org.apache.helix.guardrail.rules.LiveInstanceGuardrailRule;
 import org.apache.helix.integration.manager.MockParticipantManager;
 import org.apache.helix.integration.task.MockTask;
 import org.apache.helix.manager.zk.ZKHelixDataAccessor;
@@ -949,10 +949,10 @@ public class TestPerInstanceAccessor extends AbstractTestClass {
   }
 
   /*
-   * Guard rail coverage for the DELETE instance endpoint. STOPPABLE_CLUSTER is configured with
-   * MIN_ACTIVE_REPLICA (3) greater than NUM_REPLICA (2), so dropping any instance that hosts a
-   * partition necessarily drops that partition below its required minimum active replicas and must
-   * be blocked. None of these tests actually drop the instance, so they are non-destructive.
+   * Guard rail coverage for the DELETE instance endpoint. Every participant in STOPPABLE_CLUSTER is
+   * started and connected, so its LIVEINSTANCES znode is present and the live-instance guard rail
+   * must block (or, for dryRun, report) the drop. None of these tests actually drop the instance, so
+   * they are non-destructive.
    */
   @Test
   public void testDeleteInstanceGuardrailBlocks() throws IOException {
@@ -966,7 +966,7 @@ public class TestPerInstanceAccessor extends AbstractTestClass {
 
     JsonNode verdict = OBJECT_MAPPER.readTree(response.readEntity(String.class));
     Assert.assertFalse(verdict.get("feasible").asBoolean());
-    Assert.assertTrue(verdict.toString().contains(MinActiveReplicaGuardrailRule.RULE_ID));
+    Assert.assertTrue(verdict.toString().contains(LiveInstanceGuardrailRule.RULE_ID));
 
     // The instance must not have been dropped by a blocked request.
     Assert.assertNotNull(_configAccessor.getInstanceConfig(STOPPABLE_CLUSTER, instanceToDelete));
@@ -986,7 +986,7 @@ public class TestPerInstanceAccessor extends AbstractTestClass {
 
     JsonNode verdict = OBJECT_MAPPER.readTree(response.readEntity(String.class));
     Assert.assertFalse(verdict.get("feasible").asBoolean());
-    Assert.assertTrue(verdict.toString().contains(MinActiveReplicaGuardrailRule.RULE_ID));
+    Assert.assertTrue(verdict.toString().contains(LiveInstanceGuardrailRule.RULE_ID));
 
     Assert.assertNotNull(_configAccessor.getInstanceConfig(STOPPABLE_CLUSTER, instanceToDelete));
     System.out.println("End test :" + TestHelper.getTestMethodName());
@@ -1007,7 +1007,7 @@ public class TestPerInstanceAccessor extends AbstractTestClass {
     Assert.assertEquals(response.getStatus(), Response.Status.BAD_REQUEST.getStatusCode());
 
     String body = response.readEntity(String.class);
-    Assert.assertFalse(body.contains(MinActiveReplicaGuardrailRule.RULE_ID),
+    Assert.assertFalse(body.contains(LiveInstanceGuardrailRule.RULE_ID),
         "force=true should bypass the guard rail, not return its verdict: " + body);
     Assert.assertTrue(body.contains("is still alive"),
         "force=true should reach dropInstance, which fails on the live participant: " + body);
